ai_private_for_finance/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1

SQLAlchemy==2.0.30
pymysql==1.1.1

pydantic==2.7.4
pydantic-settings==2.3.4
python-dotenv==1.0.1

pandas==2.2.2
numpy==2.0.0

pyyaml==6.0.1
httpx==0.27.0

feedparser==6.0.11
beautifulsoup4==4.12.3
lxml==5.2.2

redis==5.0.7
qdrant-client==1.9.2
